refactor: 매핑 로직을 DomainMapper로 분리
