+++ Clipster/Clipster/Data/Repository/DefaultFolderRepository.swift
@@ -2,16 +2,21 @@ import Foundation
 
 final class DefaultFolderRepository: FolderRepository {
     private let storage: FolderStorage
+    private let mapper: DomainMapper
 
-    init(storage: FolderStorage) {
+    init(
+        storage: FolderStorage,
+        mapper: DomainMapper,
+    ) {
         self.storage = storage
+        self.mapper = mapper
     }
 
     func fetchFolder(by id: UUID) -> Result<Folder, DomainError> {
         storage.fetchFolder(by: id)
             .mapError { _ in .unknownError }
             .flatMap { entity in
-                guard let folder = toDomain(from: entity) else {
+                guard let folder = mapper.folder(from: entity) else {
                     return .failure(.unknownError)
                 }
                 return .success(folder)
@@ -22,7 +27,7 @@ final class DefaultFolderRepository: FolderRepository {
         storage.fetchTopLevelFolders()
             .mapError { _ in .unknownError }
             .flatMap { entities in
-                let folders = entities.compactMap(toDomain)
+                let folders = entities.compactMap(mapper.folder)
                 return .success(folders)
             }
     }
@@ -40,59 +45,5 @@ final class DefaultFolderRepository: FolderRepository {
     func deleteFolder(_ folder: Folder) -> Result<Void, DomainError> {
         storage.deleteFolder(folder)
             .mapError { _ in .unknownError }
-    }
-
-    private func toDomain(from entity: FolderEntity) -> Folder? {
-        guard let folders = entity.folders?.compactMap(toDomain),
-              let clips = entity.clips?.compactMap(toDomain) else {
-            return nil
-        }
-
-        return Folder(
-            id: entity.id,
-            parentFolderID: entity.parentFolder?.id,
-            title: entity.title,
-            depth: Int(entity.depth),
-            folders: folders,
-            clips: clips,
-            createdAt: entity.createdAt,
-            updatedAt: entity.updatedAt,
-            deletedAt: entity.deletedAt,
-        )
-    }
-
-    private func toDomain(from entity: ClipEntity) -> Clip? {
-        guard let folderID = entity.folder?.id,
-              let urlMetadataEntity = entity.urlMetadata,
-              let urlMetadata = toDomain(from: urlMetadataEntity) else {
-            return nil
-        }
-
-        return Clip(
-            id: entity.id,
-            folderID: folderID,
-            urlMetadata: urlMetadata,
-            memo: entity.memo,
-            lastVisitedAt: entity.lastVisitedAt,
-            createdAt: entity.createdAt,
-            updatedAt: entity.updatedAt,
-            deletedAt: entity.deletedAt,
-        )
-    }
-
-    private func toDomain(from entity: URLMetadataEntity) -> URLMetadata? {
-        guard let url = URL(string: entity.urlString),
-              let thumbnailImageURL = URL(string: entity.thumbnailImageURLString) else {
-            return nil
-        }
-
-        return URLMetadata(
-            url: url,
-            title: entity.title,
-            thumbnailImageURL: thumbnailImageURL,
-            createdAt: entity.createdAt,
-            updatedAt: entity.updatedAt,
-            deletedAt: entity.deletedAt,
-        )
     }
 }
